refactor(name): refactor names in snake case
