+++ engine/rust_engine/ftml_client_py/src/aim/parse.rs
@@ -1,3 +1,5 @@
+#![allow(non_snake_case)]
+
 use regex::Regex;
 
 #[derive(Debug, Clone)]
