+++ engine/rust_engine/ftml_client_py/src/lib.rs
@@ -1,19 +1,17 @@
-#![allow(non_snake_case)]
 #![allow(unsafe_op_in_unsafe_fn)]
-
-mod ACS;
-mod AIM;
+mod acs;
+mod aim;
 
 use pyo3::prelude::*;
 
 #[pyfunction]
 fn process_acs(py: Python<'_>, text: &str, store: &PyAny) -> PyResult<String> {
-    ACS::process_acs(py, text, store)
+    acs::process_acs(py, text, store)
 }
 
 #[pyfunction]
 fn process_aim(py: Python<'_>, text: &str, store: &PyAny) -> PyResult<String> {
-    AIM::process_aim(py, text, store)
+    aim::process_aim(py, text, store)
 }
 
 #[pymodule]
